Changing getMoney() to return an int

+++ src/main/java/com/zonkware/osbj/Player.java
@@ -24,7 +24,7 @@ public class Player {
 		return this;
 	}
 	
-	public float getMoney() {
+	public int getMoney() {
 		return money;
 	}
